Projects/Report/python_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
